techhive/backend/requirements.txt
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
flasgger==0.9.7.1
gunicorn==23.0.0
psycopg[binary]==3.2.9
